decimal to int in .DAL models

+++ MoneyStats/MoneyStats.DAL/Model/DBModel.cs
@@ -4,9 +4,9 @@ namespace MoneyStats.DAL.Model
 {
     public abstract class DBModel
     {
-        public decimal ModifiedBy { get; set; }
+        public int ModifiedBy { get; set; }
         public DateTime ModifiedDate { get; set; }
-        public decimal CreateBy { get; set; }
+        public int CreateBy { get; set; }
         public DateTime CreateDate { get; set; }
         public string State { get; set; }
